requirements.txt
Flask
pyTelegramBotAPI
requests
SQLAlchemy
pydantic
dataclasses
typing-extensions
gunicorn
python-dotenv